pyproject.toml
[project]
name = "openclaw-bot"
version = "0.1.0"
description = "Self-modifying AI coding agent that runs on your phone"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "httpx>=0.25.0",
    "openai>=1.0.0",
    "python-telegram-bot>=21.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
claw = "openclaw.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openclaw"]
